Moved dependency-resolvers to their own files & made them chainable
Instead of being done as soon as one resolve returns a file, they will
now always be chained and pass each other an object. This allows for
more flexible use of resolvers. E.g. a mapping resolver can be placed at
the start of the chain which alters the require-ed name, and likewise a
resolver can be placed at the end which will alter full paths. The later
was not possible with the old strategy.