Removed dependencyResolver timeout
Although the timeout seemed like a good idea, it is also triggered when
something else is blocking the node thread for a long time. This can
happen easily in a parallel build with large files (e.g. compiling a
large less project), so the timeout is removed all together now.