Better support for gulp-cache & gulp-remember
gulp-cache & gulp-remember work by caching the chunks that go through
them. However, if those chunks are changed further down the pipeline
they will be changed in cache as well, they're merely shallow copies. In
the previous implementation we changed the type of the `chunk.deps`
field such that it was not cachable. By now instead using the `dep.as`
property it can be served from cache.