Prevent reprocessing dependencies that are already present on chunk
This turned out to be problematic with caching: a chunk that is reused
from cache can be in multiple phases of the pipeline at once. It
therefore could happen that it's `dep` property would be reset to the
empty object just before it was about to enter the `linkDependencies`
transform, which needs the `dep` property.

We now assume that if a `dep` object is present that it is created by
yoloader and can therefore safely be reused.