Reversed variable name and package name for shimconfig
Not all packages export anything. Some packages just put something
on the window object, and others don't even expose anything at all
but register listeners/services etc. in other frameworks. If the
variable name is the key of the depends config then it's not possible
to instruct yoloader to just require the module, but not assign it
to a variable. Therefore this has now been changed such that the variable
name is the value and the package name is the key. This allows the user
to set the variable name to , which causes yoloader to not use
the exported value